(MAINT) Refactor GHA
This commit majorly overhauls the implementation of the GitHub Actions
in this repository, moving the logic for the workflows out into scripts
and adding a new `gha` module in the `.github` folder to make utility
functions and api helpers available to those scripts.

This should make maintenance and enhancement easier for the maintainers.

Finally, this commit introduces another authorization check to verify
that a pull request author may modify repo-sensitive files; namely the
workflow definitions and the PowerShell module/scripts that the
workflows depend on. This should help prevent unauthorized users from
modifying the repository dangerously.