(MAINT) Add GHA to enforce checklist usage
Prior to this commit, the PR template was something of a kitchen sink,
combining information with expectations and requirements. Now that the
expectations workflow has been added, much of this can be removed.

This commit simplifies the PR template significantly, leaving only the
summary and a slimmed-down checklist. With the remaining checklist items
being mandatory for all PRs, we can add a PR check to verify they've
been filled out.

This commit adds such a check, writing brief messages about the check
list status and failing if any are missing.